all: Format one line 'if' or 'for' statement.
Bug-AGL: SPEC-2988

+++ low-can-binding/utils/converter.hpp
@@ -23,5 +23,5 @@ class converter_t
 {
 	public:
 		static std::string to_hex(const uint8_t data[], const size_t length);
-		static void signal_to_bits_bytes(uint32_t bit_position, uint32_t bit_size, int &new_start_byte, int &new_end_byte, int &new_start_bit, int &new_end_bit);
+		static void signal_to_bits_bytes(uint32_t bit_position, uint32_t bit_size, int &new_start_byte, int &new_end_byte, uint8_t &new_start_bit, uint8_t &new_end_bit);
 };
